requirements.txt
flask
flask-cors
Flask-SQLAlchemy
flask-mysqldb
flask-login
pandas
scikit-learn
matplotlib